+++ src/Microsoft.Data.SqlClient/tests/ManualTests/DataCommon/SqlClientCustomTokenCredential.cs
@@ -18,11 +18,8 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
         private const string DEFAULT_PREFIX = "/.default";
         private const string AKVKeyName = "TestSqlClientAzureKeyVaultProvider";
 
-        private static string AKVUrl = (new Uri(DataTestUtility.AKVBaseUri, $"/keys/{AKVKeyName}")).AbsoluteUri;
-
         string _authority = "";
         string _resource = "";
-        string _akvUrl = "";
 
         public override AccessToken GetToken(TokenRequestContext requestContext, CancellationToken cancellationToken) =>
             AcquireTokenAsync().GetAwaiter().GetResult();
@@ -34,10 +31,11 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
         {
             // Added to reduce HttpClient calls.
             // For multi-user support, a better design can be implemented as needed.
-            if (_akvUrl != AKVUrl)
+            if (string.IsNullOrEmpty(_authority) || string.IsNullOrEmpty(_resource))
             {
                 using (HttpClient httpClient = new HttpClient())
                 {
+                    string AKVUrl = (new Uri(DataTestUtility.AKVBaseUri, $"/keys/{AKVKeyName}")).AbsoluteUri;
                     HttpResponseMessage response = await httpClient.GetAsync(AKVUrl);
                     string challenge = response?.Headers.WwwAuthenticate.FirstOrDefault()?.ToString();
                     string trimmedChallenge = ValidateChallenge(challenge);
@@ -69,8 +67,6 @@ namespace Microsoft.Data.SqlClient.ManualTesting.Tests
                         }
                     }
                 }
-                // Since this is a test, we only create single-instance temp cache
-                _akvUrl = AKVUrl;
             }
 
             AccessToken accessToken = await AzureActiveDirectoryAuthenticationCallback(_authority, _resource);
